+++ src/components/post/post-attachments.jsx
@@ -3,11 +3,59 @@ import ErrorBoundary from '../error-boundary';
 import ImageAttachmentsContainer from './post-attachment-image-container';
 import AudioAttachment from './post-attachment-audio';
 import GeneralAttachment from './post-attachment-general';
+import VideoAttachment from './post-attachment-video';
+
+const videoTypes = {
+  mov: 'video/quicktime',
+  mp4: 'video/mp4; codecs="avc1.42E01E"',
+  ogg: 'video/ogg; codecs="theora"',
+  webm: 'video/webm; codecs="vp8, vorbis"',
+};
+
+// find video-types which browser supports
+let video = document.createElement('video');
+const supportedVideoTypes = [];
+Object.keys(videoTypes).forEach((extension) => {
+  const mime = videoTypes[extension];
+
+  if (video.canPlayType(mime) !== '') {
+    supportedVideoTypes.push(extension);
+  }
+});
+video = null;
+
+const looksLikeAVideoFile = (attachment) => {
+  const lowercaseFileName = attachment.fileName.toLowerCase();
+
+  for (const extension of supportedVideoTypes) {
+    if (lowercaseFileName.endsWith(`.${extension}`)) {
+      return true;
+    }
+  }
+
+  return false;
+};
 
 export default (props) => {
   const attachments = props.attachments || [];
 
-  const imageAttachments = attachments.filter((attachment) => attachment.mediaType === 'image');
+  const imageAttachments = [];
+  const audioAttachments = [];
+  const videoAttachments = [];
+  const generalAttachments = [];
+
+  attachments.forEach((attachment) => {
+    if (attachment.mediaType === 'image') {
+      imageAttachments.push(attachment);
+    } else if (attachment.mediaType === 'audio') {
+      audioAttachments.push(attachment);
+    } else if (attachment.mediaType === 'general' && looksLikeAVideoFile(attachment)) {
+      videoAttachments.push(attachment);
+    } else {
+      generalAttachments.push(attachment);
+    }
+  });
+
   const imageAttachmentsContainer =
     imageAttachments.length > 0 ? (
       <ImageAttachmentsContainer
@@ -23,7 +71,6 @@ export default (props) => {
       false
     );
 
-  const audioAttachments = attachments.filter((attachment) => attachment.mediaType === 'audio');
   const audioAttachmentsNodes = audioAttachments.map((attachment) => (
     <AudioAttachment
       key={attachment.id}
@@ -39,7 +86,21 @@ export default (props) => {
       false
     );
 
-  const generalAttachments = attachments.filter((attachment) => attachment.mediaType === 'general');
+  const videoAttachmentsNodes = videoAttachments.map((attachment) => (
+    <VideoAttachment
+      key={attachment.id}
+      isEditing={props.isEditing}
+      removeAttachment={props.removeAttachment}
+      {...attachment}
+    />
+  ));
+  const videoAttachmentsContainer =
+    videoAttachments.length > 0 ? (
+      <div className="video-attachments">{videoAttachmentsNodes}</div>
+    ) : (
+      false
+    );
+
   const generalAttachmentsNodes = generalAttachments.map((attachment) => (
     <GeneralAttachment
       key={attachment.id}
@@ -60,6 +121,7 @@ export default (props) => {
       <ErrorBoundary>
         {imageAttachmentsContainer}
         {audioAttachmentsContainer}
+        {videoAttachmentsContainer}
         {generalAttachmentsContainer}
       </ErrorBoundary>
     </div>